Make pid's idea of following error match motion's
Motion generates a new commanded position each servo cycle.  It expects the
machine to move such that at the next period, it will be at that new target.
It calculates ferror accordingly (this period's feedback vs last period's
command.)

Until now, pid didn't agree about what ferror meant: it would read the
commanded and feedback positions simulatneously and call that difference the
following error.  This made it disagree with motion by the distance moved in
one period, which gave an ferror discrepancy proportional to velocity.

Unfortunately this fix has to be optional, at least in the release branch,
because it will affect existing pid tunings.