gscreen -modify cycle start method so gscreen can use it
This file was based from TOUCHY's, some of the methods are still unchanged.
The cycle start method was not used till now.
This method is currently used when the cycle-start pin is changed true.
If in paused mode then it will single step the program else start from
the beginning. I may wish to include run-from-line later