gscreen -have the velocity calculated with units
velocity will be in units per minute.
This works fine for linear units but will be wrong for
angular units when the display is not in native machine units.
AXIS has the same problem. SO for the display to always be right
one must know what type of joint is being jogged. More work to be done.