doc build: use w3c's link checker, if available
If the w3c link checker is available, run it on all our generated html
docs.  It does a better job than our checklink.py program, but it's not
available on Lucid.  (It is available on both Hardy and Precise.)

If it finds any broken links it English it fails the build.

If it finds any broken links it French (there are many) it warns but
does *not* fail the build.  When the french links are fixed we'll turn
broken links into a build failure too.