pickconfig.tcl: improve readme.ngc
The auto-generated file named readme.ngc is put in the user nc_files directory.
It can run as a simple gcode file and displays debug messages that show
the linuxcnc version and identifies important directories for the user.

With this change, the file can also run as an ngcgui compatible subfile.