doc build: make CSS and JS stuff external to the docs
Before this commit, we ran asciidoc in its default mode, where it inlines
the stylesheets and javascript inside each document.

This commit changes asciidoc to link to external style sheets and
javascript instead, so we can get by with one copy of those things.

We copy all the asciidoc CSS and JS files into our directory of generated
html docs.  asciidoc used to inline some (but not all) of those documents
in each html file.

Also move our custom stylesheet from doc/src to doc/html, where it will
be used.

The effects of this commit are:

    Copy some of asciidoc's files into our html tree.

    Slightly smaller docs: each document shrinks by about a dozen
    kilobytes (the CSS and JS files are cached by the client after the
    first fetch).

    No bogus inlining.  Inlining is discouraged by the w3c, and asciidoc
    got it wrong:

        http://www.w3.org/TR/xhtml1/#h-4.8

        http://www.w3.org/TR/xhtml-media-types/#C_4