Don't leak file descriptors or memory when abused
It's possible to call Ini_init multiple times, don't leak a new IniFile
or a file descriptor each time.

In Ini_dealloc it's still necessary to PyObjectDel self; it's defined as
OK (a no-op) to delete a NULL pointer, so the delete statement doesn't
need to be guarded by the test.