gscreen - fix preferece file option saving bug
if you getpref an option more the once, it would error often silently
but use the default value. Set it up to use raw parser instead of
interpeter - little idea what was really wrong. on ethe second
run if would try to check the value for substituion
if "%(" in value:
which cannot be in a Float value!