tests: try to work around a timing problem in t0 tests
It seems like if the UI-to-Task NML queues get too long, MDI commands
can get dropped.

linuxcncrsh has a mode, "set_wait=done", which is supposed to make it wait
for each issued command to finish before issuing the next, but it hangs if
the command fails (as many of the commands in this test do, deliberately).

This commit adds explicit, non-modal waits after some of the known-good
MDI commands.

I have been unable to reproduce the bug in my testing after this commit,
but I was never able to reproduce it very reliably before this commit
either...