expose linux's module_param() before using it
This commit changes rtapi.h to include <linux/module.h> in all our rtapi
code, but to never bother including <linux/param.h>.

linux/param.h has nothing to do with module parameters.  Including it
does us no good.  The functions that rtapi.h tries to use are defined
(indirectly) in linux/module.h, so that's the one that we need to include.

The hostmot2 pins.c file is the only one we have (i think) that uses
module_param(), via RTAPI_MP_INT(), without explicitly including
linux/module.h.

This works on 2.6 kernels because (i think) one of the files that pins.c
happens to include eventually ends up including linux/module.h.  It does
not work on 3.x kernels because this presumed indirect inclusion does
not happen.  So in 3.x, we need to explicitly include linux/module.h to
use module params.

The proper place to do that is in the rtapi.h header right before it
tries to use module_param().

I'm making this change in 2.5 instead of master since the bug is present
in 2.5 too, even though it is only being triggered in master.