Fix the top of the maxvel slider being too low
On jensor's machine, he had a Z faster than X, but the maxvel slider
defaulted to having a maximum setting that matched X.  This caused
a lot of confusion.  The docs say this setting should be in DISPLAY
but for this purpose only TRAJ was being searched.

This change allows either location, just like for jog speed.