add a a halui jogging test
This test checks jogging via halui.

The test stands up a full simulated linuxcnc instance, including halui
and a custom [DISPLAY]DISPLAY "ui" written in python.

This python "UI" uses the linuxcnc python module and the hal python
module.  The python script creates some HAL pins, which are netted to
halui (for commanding) and motion (for feedback).

The python UI pokes halui's pins to select joints and to jog in the plus
and minus directions, and monitors motion's axis.N pins to verify that
the correct jogs happened.