Revert "queued MDI:  support in task & Axis"
This reverts parts of commit 55d93a8fea53bb5963af47143e44d00d58a664bd:

The new EMC_TASK_PLAN_EXECUTE_INTERNAL_TYPE NML message is reverted (it is
not needed and complicates things).

The new (and incorrect) MDI queueing logic is reverted.

The mdi input queue in task remains, though all use of it is reverted
(because it is buggy).  The queuedMDIcommands field in EMC_TASK_STAT
remains.  The changes to Axis to check the mdi queue len remain.