build: disable -funsafe-math-optimizations
This flag causes miscompilation on i386 and x86_64 using gcc 4.7.2-5
on debian wheezy due to upstream bug 56125
    http://gcc.gnu.org/bugzilla/show_bug.cgi?id=56125

there is at least one report of a running realtime system on amd64
requiring this flag (that's why it's there!), though not one
linuxcnc.org is shipping to users.  I've communicated with the
original reporter and if this is needed in his system we'll figure
out a configure-time test.