build: revisit x86_64 build flags yet again
Motivation to add -msse conditionally:

On x86_64, the kernel specifies -mno-sse, with the intention of detecting
accidental use of FP in kernel space.

gcc -m64 -mno-sse rejects the following code, because the ABI places the
return value of double sin(double) in the SSE register %xmm0 (and/or
because the argument :
    extern double sin(double);
    double d;
    void fn() {
        d = sin(d);
    }

rtai for x86_64 is supposed to put -msse in $(rtai-config
--module-cflags), but in rtai-modules-2.6.24-16-rtai 3.6.1-linuxcnc.2
(ubuntu hardy amd64) this is not the case.

Detect this and add -msse by hand if it is not there.  Also, take care
to add it only in kernel space; we're not aware of any reason to have to
add it in userspace.

Motivation to remove -ffast-math, -f[no-]unsafe-math-optimizations:

gcc 4.7.2-5 on debian wheezy miscompiles
    double val;

    void fn(double dot, double radius) {
        val = dot / (radius * radius);
    }
under -Os -ffast-math.  -ffast-math implies -funsafe-math-optimizations,
so specifying it was redundant.  In 2011 I thought I had a report that
-ffast-math -funsafe-math-optimizations was required on one user's amd64
setup, but now that same user reports he has to drop those flags to get
correct behavor, so as far as I know we have no users who need the
-ffast-math flag. (this bug is upstream and fixed:
http://gcc.gnu.org/bugzilla/show_bug.cgi?id=56125 but the fix looks
unlikely to make it into Debian 7.0's initial release)