interp: add __delete__ interpreter shutdown hook to Interp::exit()
this mirrors the Interp __init__ hook (see section 9.3 remapping manual)
intended use case is to enable any user-defined initialisation of state,
and saving of state on shutdown as far as this is not covered by
Interp::save_parameters().

This patch enables the __delete__ hook only for the milltask instance
of interp, where Interp::exit() is properly called.

TBD: how to call Interp::exit() on gcodemodule.cc shutdown