task: always allow coolant control
Manual control of flood & mist coolant used to be disabled while Task
was in Auto mode and Interp was Running or Waiting.  But it was allowed
while Task was in Auto mode and Interp was Idle or Paused.

This commit enables coolant control in all Interp modes while Task is
in Auto.