task: always allow coolant & lube control
Manual control of coolant (both flood and mist) and lube used to be
disabled while Task was in Auto mode and Interp was Running or Waiting.
But it was allowed while Task was in Auto mode and Interp was Idle
or Paused.

This commit enables coolant and lube control in all Interp modes while
Task is in Auto.

I asked if this was change was right and Andy Pugh told me to push it.