remove some dead code in halrmt
This particular chunk of code looks like a place holder for future work
that never materialized: do_loadusr_cmd() is not available in halrmt,
and without that function call this code has no effect.  It's been this
way since 2008.

Since it's causing a compiler warning i'm disabling it with "#if 0"
for now.