genserkins: reduce stack usage
Normally I like limiting the scope of local variables as much as
possible, but in this case two adjacent scopes both instanciate a large
anonymous struct, and the compiler doesn't realize they're identical
and have non-overlapping use, and it puts both of them on the stack.
This results in a very large stack frame for this function.

This commit moves the declaration of the large structure out into the
enclosing scope, and both of the two nested scopes use the shared struct.