docs: move linuxcnc.1.in out of the manpage tree
The man/man? directories are for manpages, not for the source code
for manpages.

mandb complains about any non-manpage files it finds in there.

We have one non-manpage file in those directories: man1/linuxcnc.1.in,
which is turned into linuxcnc.1 by config.status (called by make),
to embed the search path for .ini files into the manpage.

This commit moves linuxcnc.1.in into the docs/src tree, so as not to
pollute the manpage directory tree.