sanify shmdrv build
Create rtapi/shmdrv/Submakefile and move relevant portions of
rtapi/Submakefile and rtapi/shmdrv/Makefile to it.

Patch into larger build system.

FIXME: The ub build system needs to provide CONFIG_SHMDRV=m if it is
decided to build shmdrv, and making modules needs to call kbuild for
just shmdrv on userland thread style builds.

FIXME: The ub and this module has funny implications for userland
threads.  If every set of kernel headers is detected and this module
built for it so that the module is available to userland threads for
all installed kernels, this will need to be packaged in a separate
kernel-specific package, much like kernel threads modules will be.
Does this make sense?  Should it only be built for kernels that
support kernel threads systems, i.e. RTAI and Xenomai?