shebang python, not python2
All the platforms i've checked (Hardy, Lucid, Precise, and Wheezy) have
a symlink named python pointing at a python2 interpreter, but not all
have a symlink named python2.

If/when python3 becomes widely used we'll have to solve this in a more
clever way, but for now it works fine.