configure: require libmodbus3
By default, configure will now look for libmodbus (installed from a
separate deb, provided by Ubuntu for Precise and by the linuxcnc.org
debian archive for Hardy and Lucid) and error out if it can't be found.

The user can disable this check (and disable compilation of all drivers
that use libmodbus) by configuring with "--without-libmodbus".