gladevcp fix calculator widget math error
numbers starting with a 0 such as 0123+1 would be
converted to octal instead of decimal (so above = 84)
also dividing a whole number into a faction would not
show the decimal portion (1/2 would not = .5)
now we split the string up and qualify all the parts.
converting the numbers to floats first. Then we put it back
together and eval() it.

This bug affected Gscreen's run-from-line badly and certain
division calculations as above.