gladevcp -add search, undo and redo functionality to hal_sourceview
To facilitate small edits of gcode.
Right now the actual 'editting mode' and controls for said fuctions are
done outside hal_sourceview.
It is possible to get the linuxcnc Gcode and the sourceview buffer Gcode
out of synch if you forget to save the result of the edit.
If the integrator was forced to set sourceview into an edit mode, it would
be easier to track this. there is a 'has-been-modified' flag to check.