fix a config reporting bug in gs2_vfd
This commit fixes up a bug I introduced in
fc9e19b414c409a153c5651f3a4466a87dd0622c.

That commit intended (among other things) to change the order of some
of the configuration information that gs2_vfd reports at startup.  But I
got it wrong: I reversed the order of a pair of conversion specifiers in
the printf format string without also reversing the order of the later
arguments passed in.

This commit reverses order the variables passed in to match the format
string.