touchy: allow wheel scrolling of program start point
The start point can be moved by 1/10/100 lines per click.  On a
100CPR encoder this means you can go 10000 start-points per
revolution, so this should be good enough for part programs under a
few hundred thousand lines.  (if we discover that longer programs
need to be scrolled, 10/100/1000 increments + the existing onscreen
1-line buttons might suffice)

So that the width of the increment labels does not change, a
(minimum) char width must be est on them.  Set it on the axis
selection widgets too, so that they are the same width as their
siblings.

Modify widget state management as necessary.