gladevcp -change the layout and control functions of offsetpage_widget
for consistancy have the axes in the colomns and the offsets in rows.
You can now edit any user system.
You can add descriptions of the systems.
The current system's text is in red, others are black.
you can hide or show any colomn or row.
different highlighting options

Gscreen is changed to not error - rows and colums are switched now.