gladevcp -fix offsetpage widget out of index error
When using entering data into a config that didnt start with
X axis and wasn't continous to the last axis (XYZABC) would
give an error.
This was because we excluded the extra axes letters in the variable.
The variable is used to convert axes letter to number.
now the variable always has all the axes so the lookup number is right.