test: add a test of M61
This test uses a full LinuxCNC setup and a custom python test-ui.

It commands M61 and verifies both that the IO tool-change pins don't act,
and that the tool-related interp parameters reflect the requested tool
when it's done.

This test fails on v2.5.3-9-g5f7693a but passes after the m61-fix commits
are applied.

Thanks to Norbert (nieson) for reporting this bug.