m61 test: don't throw exceptions in LinuxcncControl.g()
By default the LinuxcncControl class throws an exception if it gets an
unexpected message from linuxcnc while waiting for its gcode to run.
This would cause spurious test failures when we got the "Unexpected
realtime delay" message on the laggy VMs in the buildbot.

This commit disables these exceptions, the messages just get printed to
stdout now.