interp: avoid violating Python runtime contract
Before, Interp::_setup would be destroyed after the program exited,
which is generally after Py_Finalize is called.  But the
destruction of Interp::setup implicitly destroys its member boost::python::
object pythis, leading to Python API calls including to subtype_dealloc,
which causes routine crashes when axis exits on Python 2.7.5-4
(Debian Unstable)

Now, by the use of references and memory allocated with operator new,
the destructor is never called for _setup; we just leak the memory instead.
This feels like a band-aid fix for the real problem, but Michael H. and I
agreed in an irc discussion that this fix is an improvement for now and
won't be much of an obstruction when fixing the static interpreter data
later on.