Revert "interp: avoid violating Python runtime contract"
This reverts commit 33380c4b144b5508717209f634bc2e98aaba17ba.

This "fix" fails on at least some machines.  I think the flaw in my analysis
has to do with moving _setup from all-bits-zero initialized storage to new'd
storage, so all the scalar fields are undefined, but I am not 100% confident
this is the right explanation.

In any case, the cure was apparently worse than the disease, and we'll have to
wait for MAH's more elegant and thorough fix.