Stripped out remaining PmPose and PmLine references in TP.
An alternative to PmLine was added in PmCartLine, which strips out all
of the rotation information. Because each axis is treated as orthogonal
and linear as far as trajectory planning, there is no use for an
additional rotation term. More complex kinematics are dealt with
external to TP, so there is no need for the extra space and processing.