docs: Don't explicitly request a dblatex encoding on newer(?) asciidoc
On Debian jessie (and wheezy) this leads to problems e.g., when
non-ACSII characters are used in code listings.  However, other
systems (e.g., Ubuntu Precise), this flag is required.  Skip
specifying the flag when a simple asciidoc source file can't be
built into a pdf document when the flag is specified.