Potential fix for "sawtooth" effect
Added a candidate check for blend arcs to make sure that we don't
shrink the previous segment too much. If we can't entirely eliminate
the previous segment, then it's better to shrink the blend arc so that
the line segment is long enough for TP to land on at least once.