Fixed parabolic blending bug introduced previously
Parabolic blending was crashing out due to an improper estimate of being
on the final deceleration. The final decel calculation assumed that we
were on the final deceleration if the ideal new velocity was close to
what we could actually reach based on machine limits. This should only
be true when we are decelerating to a stop.
TODO: what exactly changed that made this assumption not work anymore?