axis: drop seticon hack
Since tk8.4.8, [wm iconphoto] has been able to set photo images as
window icons, so the seticon hack is not necessary.  Furthermore, it was
causing crashes on some systems and ineffective on others.

Also, the Img package (which can load png images) has been required for
the latencyhistogram script, so this is not a new requirement.