Added a simple cutoff to disable smoothing
The price of velocity smoothing is that it caps the velocity of the
previous and next segments. Adding a minimum blend ratio skips smoothing
for segments where the arc is short compared to the lines. This lets the
line segments accelerate to the full requested velocity, rather than
being capped at the arc's velocity.