Added final velocity override for intermediate conditions
Partially addresses issues seen during wait-on-spindle moves. if the
trajectory component was changed to a stop condition, the final velocity
was used anyway. Now, any non-tangent behavior returns a 0 final
velocity. There may be edge cases here still, and this really suggests a
bit of a rethink of this structure. The band-aids are getting deep...