Performance improvements
Now that split-cycle tangents are working well, it was time to revisit
the restrictions placed on blend arc size previously to identify any
performance bottlenecks.
* Removing the blend_ratio limitation completely led to inconsistent arc
  sizes
* Removing smoothing entirely (unsurprisingly led to velocity peaks)

The modified version here strikes a balance that leads to faster, smooth
motion on "spiral-in.ngc". There are subtle changes in how the blend
ratio is calculated (basically assuming that we want the same ratio
between the arc and both the prev. and next line).

Also, smoothing with the next segment turned out to be unnecessary. The
reason it helped before is that optimization was ending prematurely,
which introduced small ripples in velocity.