stepconf -fix a regression - spindle-at-speed not set
If the right combination of feedback and pyVCP was selected
the spindle at speed signal was not set true.
If you have feedback a pyvcp spped indicator and
select the use spindle-at-speed the near component
will be used to compare commanded vrs feedback of the spindle.
any other combo forces the spindle-at-speed always true.