deb: unconditionally build-depend on imagemagick
Before this commit only the Ubuntu distros (Hardy and newer)
build-dependend on imagemagick, and every other platform tried to build
without it.  But the docs need the 'convert' program from imagemagick
to build, so make that build-dependency unconditional.

Maybe the previous wonky behavior was a left-over from when we used
to support even older distros, that didn't ship with imagemagick?
Or maybe it came about by accident during the lyx-to-asciidoc switch?
I really don't know.