deb: make dependency on tcl/tk unversioned
Hardy through Precise, and Wheezy, all provide unversioned tcl/tk packages
(including the -dev packages), in addition to the versioned ones.

We've always dependend on the versioned ones (eg tcl8.5-dev), at the
cost of some configure-time hassle, and i think this is not needed.
Maybe it was needed in the distant pre-Hardy past, but we can get away
without jumping through this hoop nowdays.