mesa-hostmot2 driver:  Remove int->double conversions
Fixes 'Unknown symbol __fixunsdfdi' on 32-bit gcc-4.[34].x kthreads

For 64-bit arithmetic, gcc provides helpers on 32-bit architectures.
The __fixunsdfdi routine converts a double to u64.  For some versions
of gcc, this symbol is left for the linker to resolve from libgcc.a.
Kbuild does not link from libgcc.a, so these must be worked around.

The sserial.c example casts the double pin->fullscale to a float so
the expression is evaluated as a float, and then casts the float to
u32, avoiding double and u64 types in the conversion.

The watchdog.c example replaces the floating point calculation,
presumably meant to avoid loss of precision in an integer divide, by
doing the multiplication in 64-bit space first, then using the
kernel-supplied do_div() macro for the 64-bit integer division.

Also see these links for where these issues were addressed previously:

http://permalink.gmane.org/gmane.linux.distributions.emc.user/18718
http://permalink.gmane.org/gmane.linux.distributions.emc.devel/3663

(There are several more places in components/lcd.c parse_fmt() that
need these fixes; this patch at least the hostmot2 driver and unit
tests working.)