runtests script:  run with 'bash -e'
Without '-e', runtests can exit 0 when tests didn't actually run.

This was observed when the /tmp filesystem became full, and
$TMPDIR/alltests couldn't be written.

When '$?' is checked, this is handling a possible non-success exit
from the previous command.  Modify each such command so that the exit
code is collected but the command doesn't result in the script
exiting.