Increased blend radius for short segments
For nearly-tangent blends, a minimum-size the blend arc can be very short and still
satisfy the minimum radius to be within limits. However, this leads to
sudden short pulses of acceleration during corners, which could
introduce vibrations.

The modification here grows the radius until the amount of each segment
consumed is the same as the equivalent parabolic blend.