stepconf, pncconf: disable check_rt() for now
att: Chris Morley

the logic of testing a compiled-in symbol from halmodule.cc which depends on
sim vs rt does not make sense anymore.

What we have is a current thread flavor, and a list of flavors possibly
supported by a build.

The most sensible option IMO is to test for posix threads currently running
and isssue a warning. I'll see if I can add a flavor attribute to  halmodule.cc .

linuxcncconfig.py isnt helpful here because it carries compile-time
information.