xhc-hb04: support different stepsize_sequences
The default stepsize sequence: 0.001,0.010,0.100,1.000 (after multiplication
by .001 unit) is often too big for inch-based machines:
  a) in 'normal' jogging (the default), the larger stepsizes will allow
     the machine to continue traveling after the wheel stops and this may
     be unexpected by the user.
  b) the alternate 'vel-mode' jogging addresses a) (since the machine stops
     when the wheel stops), but the larger stepsizes may be too big for an
     inch-based machine with typical velocity/acceleration limits.
     Example (sim/axis/axis.ini):
         vmax=1.2,amax=20, max_move = 0.5*vmax**2/amax = 0.036
     This 0.036 inch limit results in unexpected movement amounts for
     the larger stepsizes in the sequence (0.100,1.000)

This commit defines the default stepsize_sequence: 1,10,100,1000 (no change)
and allows specification of an addtional sequence: 1,5,10,20.  With this
sequence, the inch step sizes are 0.001,0.005,0.010,0.020 and will result in
expected movement when the wheel is moved one click at a time.

(Note also that the ini setting [XHC_HB04_CONFIG]scale can supply an individual
scale multiplier for each axis)

The stepsize sequence is specified with the -s option:
 -s: step sequence:
     1: 1,10,100,1000 (default)
     2: 1,5,10,20

Additional sequences can be added in the future; the sequences are limited
by the display codes available for the xhc-hb04.

also:
   updated README and .txt files
   xhc-hb04-layout2_mm.ini: new file for 9axis mm example
   monitor_xhc-hb04.tcl: fail more gracefully if xhc-hb04 comp didn't load