Rework debian/configure
(Apologies to Seb for trampling his hard work!)

The previous debian/configure script disallowed module packages from
building in some cases, even when all dependencies are satisfied:

- Building under a non-RT kernel, or an RT kernel with no headers
  installed:  script exits

- Under a supported RT kernel, flavors incompatible are not built

In addition, it supported a complicated legacy substitutions system of
which much could be replaced by dh_gencontrol and substvars.

This patch replaces debian/configure with a simpler shell script that
performs only one substitution in Build-Depends:, and replaces other
substitutions with substvars.  It also detects thread flavor deps
already installed on the system, and those detected are handled in the
same way as before, by appending control file fragments.

This configure script does lack the configurability of the previous
system: it assumes that posix and rtpreempt threads should always be
built, and also any other thread styles for which deps are installed.
However, this behavior should be easily changed.

In addition, this patch removes other legacy templates for simplicity,
since there was no use of templating in debian/*.files and
debian/rules is a Makefile.

Conflicts:
	debian/configure