interp: don't close log file if it's stderr
This fixes a problem that Dewey noted in private e-mail
where "puts stderr" logging was broken after loading an ngc file.
The combination of a change which used stderr as the interpreter's
default logging destination, and a change which caused interpreter objects
to be created and destroyed within a singla AXIS session led to this
problem.