master doesnt build on hardy right now
The problem is the xhc-hb04 driver, which uses libusb-1.0-0-dev, which
was added to Ubuntu in Lucid.

This commit disables Hardy builds on the buildbot, but a better fix is
probably to just disable the xhc-hb04 driver on Hardy.  That requires
some involved changes to both debian/configure and src/configure.