gmoccapy_1_0_5 - cleaned up the gmoccapy/config/dir
I cleaned up in the config dir, so we only have one postgui file,
I moved all other needed hal connections to own hal files and that
ones are loaded with the embedded tab commands.

To all Admins: please do not delete or ignore the

gmoccapy_tool_sensor.para

file. It is a preference file and has on purpose the ending para to
not being ignort from git. Without this file the user will miss
initial settings and the simulation may not run!

Norbert