posix/rtapi_get_clocks: fall back onto clock_gettime() if no TSC available
there might be a TSC equivalent on ARM, but it cannot be used without
enabling in kernel mode. Use this until we know better, or can provide
an option to use a hardware TSC.

Depending on how clock_gettime() is implemented, this may not be a
delay factor any more; glibc on at least x86 and amd64 uses a vDSO call
for this.