Troubleshooting canon
Reverted to 1-indexing, but this didn't solve the problem. It turns out
that XY rotation is to blame for choosing the wrong axis start / end
points, which had butchered the axis_len calculation. This current state
may not be right, but it runs on a few test programs so far.