halcmd: Better detection of some arrow-related typos
This fixes a confusing result where the command

	net name firstpin <=secondpin => thirdpin

(note the missing space) would connect firstpin and thirdpin, but leave
secondpin unconnected, and would not give an error.

With this change, only the three expected arrows: =>, <=, <=> (see man halcmd)
are accepted, and only if they are a separate word that matches exactly; this
causes the net command above to fail with:

	Pin '<=secondpin' does not exist

This will fail unexpected arrows like ===> that were previously accepted but
never documented as accepted.