biquad: fix a crash-when-enabled bug
This fixes a bug in the biquad comp that would cause a crash with a NULL
pointer dereference immediately upon the user setting the .enable HAL
pin to True.  Thanks to forum user gaston48 for the bug report.

The biquad comp exports a hal function called "_" in the .comp and .c
files, called "biquad.0" in hal.  This function calls a function internal
to the biquad component (not exported to hal, not handled by comp(1))
called "Biquad_CalcCoeffs".  The Biquad_CalcCoeffs function accesses the
comp instance's hal pins and params, but this behavior is not supported
by the comp(1) tool.

Only functions handled by the comp tool can access the comp instance's hal
pins and params.  That means functions declared with a .comp "function"
declaration, and the EXTRA_SETUP() function.

This commit moves the Biquad_CalcCoeffs code from its separate function
into the main "_" aka "biquad.0" function, this avoiding the disallowed
"access hal from a called function" problem.