comp: remove an unused and confusing global variable
This commit removes an unused global variable named __comp_state, from
.c files generated by the comp(1) tool.  The global was never set,
so it never served any actual purpose.

The functions declared by "function" declarations in the .comp file,
and the optional EXTRA_SETUP() function, all take an argument which is
also named __comp_state.  Thus, in these functions, the global was masked
by the function argument of the same name.  The __comp_state variable
is used by the convenience defines to access the hal pins and params of
the component instance.  The functions all worked fine and accessed the
hal pins via their argument, just like they should.

The problem was that the global made the convenience defines compile-time
valid everywhere in the .comp, but at run-time any non-"function",
non-EXTRA_SETUP code would access the uninitialized global and crash.

We're better off without it.