docs: fix up some comp docs that had rotted a bit
The docs referred to an old name of an internal data structure.  The code
changed in a7a7fd39a but i forgot to update the docs.

The new name includes the underscore character "_", which is meaningful
to asciidoc.  This commit changes the markup delineating the name from
single-tick "'" to backtick "`", to tell asciidoc to pass through the
text without further substitution.  This also changes the rendering
from italic to monospace, which is arguably better anyway (though it's
somewhat inconsistent with the surrounding markup).

This commit also fixes a markup syntax blooper in the FOR_ALL_INSTS()
documentation.