script : fix of bug in postprocess script w.r.t. recharging
When the post processing script finds a retract anchor, it replaces
that with the M commands for retracting. Meaning setting the time
of the retract (time = 0 means 0 seconds retract disconnecting of
the extrudion with the motion). in some cases the anchor for that
in the slic3r generated code is off, resulting in the time being set
after the retract (so no retract) and a precharge then gives too much