This is useless.
This hasn't been updated for many years because it's impossible to do:
with the tiny bit of documentation sprinkled among an auto-generated
tree it's impossible to update the tree even if you cared to.  But you
don't care to, because maintaining a file that says docs are in the
directory named docs, and scripts are in the directory named scripts, is
stupid.