docs: fix some anchor names in ini_config
The [AXIS_n] and [EMCIO] sections used that bogus asciidoc syntax where
the anchor name went on the same line as the section heading words.
This makes some weird version of the HTML rendering of the section
heading appear in the anchor name.

It used to look like this:

    ini_config.html#_axis__lt_num_gt_section_a_id_sub_axis_section_a

This commit fixes the syntax so the anchor appears on the line before
the section heading, which makes everything copasetic.

Now it looks like this:

    ini_config.html#sub:AXIS-section