component : update config with joint to cartesian position
this configuration for the lineardelta uses a component that
translates the joint positions to the cartesian positions. Since
the JA4 branch is not merged yet this is a temporary solution in
getting the correct nozzle velocity for use with velocity
controlled extrusion.

install:
comp --install lineardeltajointscartesian.comp

This component is directly taken from the lineardelta kinematics.