hostmot2: fix 'set but unused' compiler diagnostics
In these cases, 'r' was intended to be used to hold the negative
errno value in the case of a failure.  Ensure that r starts with a
negative errno value (Arbitrarily, -EINVAL) and propagate it back to the
caller.

In some cases, this means an already detected failure will actually
be propagated up to the caller.  For instance, hm2_bspi_parse_md would
return success even when it failed to allocate memory.

In a few cases, a return of -1 was also changed to -EINVAL, in order to
obey the usual 'return value is negative errno value' idiom.