Let the splash screen run with default varfile
On a first startup, there are no G54 offsets, so the splash screen
will not run without touch-off.  This is confusing for new folks.

Having +Z allowed is the least bad solution for this I think,
because people do actually mill the splash screen into things,
so it should continue to have top of workpiece as zero.

Shipping and installing a nondefault varfile is the only other
solution I can think of, and that's nasty.