Container: prune fields for RT usage
the nanopb bindings will not generate, or understand fields which are
tagged with [(nanopb).type = FT_IGNORE]. This helps bring down the size
of the pb_Message struct.

The goal is to have serialisation/deserialisation happen at the RT boundary
in userland and only ship the nanopb-generated pbMessage struct over the
ringbuffer, which is about 6.5K without those options.