docs: remove some broken asciidoc markup
(Variables) looks like an typoed attempt at creating an asciidoc anchor,
but it's not valid asciidoc anchor syntax so it doesnt work, and it
passes through into the generated docs, where it makes no sense.

None of the other asciidocs tries to use an anchor named Variables,
so just remove it.