stepconf: non-fatal check for duplicate IDs
We discovered in 2014 that duplicated IDs in different glade files don't
behave consistently when those IDs are referred to from Python code.

However, some duplicate IDs seem to be OK, such as when the ID is never
referred to from Python code.

With this change, there will be a warning about *any* duplicated id in
stepconf's glade files; this produces a lot of warnings at the moment.

This should be expanded to the other parts of LinuxCNC that use glade
such as pncconf.

Related:
http://stackoverflow.com/questions/1378582/using-multiple-glade-files-gtkbuiler#1380024
http://pygtk-id.blogspot.com/2009/06/replace-duplicate-widget-name-in-glade.html