latency-histogram: Use getconf for ncpus
The old /proc-parsing code is wrong on ARM systems, for whatever reason.
This is right everywhere I checked.

getconf is in libc-bin, so it should be available on all Linux.  (In fact,
getconf is specified by POSIX, but _NPROCESSORS_ONLN is a Linux extension.
Still, this is no less portable than grepping /proc)