webtalk: integrate plugin build, add ZWS1.0 support
Websockets do not map well onto zeroMQ framing, so a policy is needed how
to translate messages depending on use case. this is what policy plugins
are for; those can be selected on a per-connection basis with the
policy=<name> parameter.

See webtalk_example_policy.cc and the Submakefile how to build a
custom relaying policy.

Plugins are loaded like so (multiple plugins possible):

webtalk  --plugin ../lib/webtalk/plugins/zws.o

The main policy used right now is 'json', which is built-in.

ZWS is a Websockets framing protocol to transport zeroMQ multiframe messages
and is better suited longer term, see https://github.com/somdoron/rfc/blob/master/spec_39.txt.
ZWS will be enabled by the policy=zws URI parameter.

It is implemented as a plugin for webtalk.

To enable the zws and pbzws, configure with --enable-dev .